% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.homo.R
\name{mle.homo}
\alias{mle.homo}
\title{Ekbohm's MLE-Based Test under Homoscedasticity}
\usage{
mle.homo(
  data,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{a 2-by-n or n-by-2 partially matched pairs samples data frame or
matrix.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify
just the initial letter.}

\item{mu}{a number indicating the true difference in means.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A list with class "htest" containing the following components:
  \item{\code{statistic}}{the value of the test statistic.}
  \item{\code{parameter}}{the degrees of freedom of the test statistic or
  NA.} \item{\code{p.value}}{the p-value of the test.}
  \item{\code{conf.int}}{a confidence interval for the mean appropriate to
  the specified alternative hypothesis.} \item{\code{estimate}}{the estimated
  mean or difference in means.} \item{\code{null.value}}{the specified
  hypothesized value of the mean or mean difference.}
  \item{\code{stderr}}{the standard error of the mean (difference).}
  \item{\code{alternative}}{a character string describing the alternative
  hypothesis.} \item{\code{method}}{a character string indicating what type
  of test was performed.} \item{\code{data.name}}{a character string giving
  the name(s) of the data.}
}
\description{
\code{mle.homo} is used to perform Ekbohm's MLE-based test
  under homoscedasticity for partially matched samples, specified by giving
  a data frame or matrix, testing hypothesis, true mean and confidence level.
}
\details{
Whether the dataset input is partially matched samples will be
  checked. If not, appropriate hypothesis tests, such as one and two sample
  t-tests, will be performed instead.
}
\examples{
# pm is a sample dataset for the PMLi package

# Ekbohm's MLE-based test under homoscedasticity formula interface
mle.homo(pm, "less", conf.level = 0.99)

# p-value of Ekbohm's MLE-based test under homoscedasticity
p.value1 <- mle.homo(pm)$p.value
}
\references{
Kuan P F, Huang B. A simple and robust method for partially
  matched samples using the p-values pooling approach. \emph{Statistics in
  medicine}. 2013; 32(19): 3247-3259.

  Ekbohm G. On comparing means in the paired case with incomplete data on
  both responses. \emph{Biometrika}. 1976; 63(2):299-304.
}
\seealso{
\code{\link[stats]{t.test}} for the one and two-sample t-tests.

  \code{\link[stats]{var.test}} for the F test to compare two variances.

  \code{\link[stats]{shapiro.test}} for the Shapiro-Wilk test of normality.

  \code{\link[stats]{wilcox.test}} for the one and two-sample Wilcoxon tests.

  \code{\link{weighted.z}}, \code{\link{modified.t}},
  \code{\link{corrected.z}}, and \code{\link{mle.hetero}} for other
  statistical approaches for partially matched samples.
}
\author{
Kai Li \email{kai.li@stonybrook.edu}
}
