% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm}
\alias{pm}
\title{PMLi Package Sample Dataset}
\format{
A data frame with 192 rows and 2 variables:
\describe{
  \item{\code{Sample1}}{Sample 1 of the matched samples}
  \item{\code{Sample2}}{Sample 2 of the matched samples}
}
}
\source{
\url{http://www.ams.sunysb.edu/~pfkuan/Teaching/AMS597/Data/d_logret_6stocks.txt}
}
\usage{
pm
}
\description{
A synthetic dataset containing partially matched samples of Sample 1 and 
Sample 2
}
\references{
Kuan P F, Huang B. A simple and robust method for partially
matched samples using the p-values pooling approach. \emph{Statistics in 
medicine}. 2013; 32(19): 3247-3259.
}
\author{
Kai Li \email{kai.li@stonybrook.edu}
}
\keyword{dataset}
