% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.z.R
\name{weighted.z}
\alias{weighted.z}
\title{Liptak's Weighted Z-Test}
\usage{
weighted.z(
  data,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  w1 = NULL,
  w2 = NULL,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{a 2-by-n or n-by-2 partially matched samples data frame or
matrix.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify
just the initial letter.}

\item{mu}{a number indicating the true difference in means.}

\item{w1}{a number indicating the weight for the first p-value. If the weight
is not inputted, a default weight is assigned to to be the square root of
two times the paired sample size.}

\item{w2}{a number indicating the weight for the second p-value. If the
weight is not inputted, a default weight is assigned to to be the square
root of the sum of the two independent samples size.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A list with class "htest" containing the following components:
  \item{\code{statistic}}{the value of the test statistic.}
  \item{\code{parameter}}{the degrees of freedom of the test statistic or
  NA.} \item{\code{p.value}}{the p-value of the test.}
  \item{\code{conf.int}}{a confidence interval for the mean appropriate to
  the specified alternative hypothesis.} \item{\code{estimate}}{the estimated
  mean or difference in means.} \item{\code{null.value}}{the specified
  hypothesized value of the mean or mean difference.}
  \item{\code{stderr}}{the standard error of the mean (difference).}
  \item{\code{alternative}}{a character string describing the alternative
  hypothesis.} \item{\code{method}}{a character string indicating what type
  of test was performed.} \item{\code{data.name}}{a character string giving
  the name(s) of the data.}
}
\description{
\code{weighted.z} is used to perform Liptak's weighted Z-test
  for partially matched samples, specified by giving a data frame or matrix,
  testing hypothesis, true mean, the corresponding weights in Liptak's
  weighted Z-test and confidence level.
}
\details{
Whether the dataset input is partially matched samples will be
  checked. If not, appropriate hypothesis tests, such as one and two sample
  t-tests, will be performed instead.
}
\examples{
# pm is a sample dataset for the PMLi package

# Liptak's weighted Z-test formula interface
weighted.z(pm, "less", conf.level = 0.99)

# p-value of Liptak's weighted Z-test
p.value1 <- weighted.z(pm)$p.value
}
\references{
Kuan P F, Huang B. A simple and robust method for partially
  matched samples using the p-values pooling approach. \emph{Statistics in
  medicine}. 2013; 32(19): 3247-3259.

  Liptak T. On the combination of independent tests. \emph{Magyar Tudom
  Aanyos Akad Aemia Matematikai Kutat Ao Intezetenek Kozlemenyei}. 1958;
  3:171-197.
}
\seealso{
\code{\link[stats]{t.test}} for the one and two-sample t-tests.

  \code{\link[stats]{var.test}} for the F test to compare two variances.

  \code{\link[stats]{shapiro.test}} for the Shapiro-Wilk test of normality.

  \code{\link[stats]{wilcox.test}} for the one and two-sample Wilcoxon tests.

  \code{\link{modified.t}}, \code{\link{corrected.z}},
  \code{\link{mle.hetero}}, and \code{\link{mle.homo}} for other statistical
  approaches for partially matched samples.
}
\author{
Kai Li \email{kai.li@stonybrook.edu}
}
